refactor: useInput 커스텀 훅 분리
